Fixed pixel precision error
The window size was resulting in incorrect pixel placement in some areas
because if the window size is not a multiple of the render target, the
target can't be perfectly projected onto the screen. A future solution
might draw into a subregion of the window, but currently it just decides
on the best size of the window. The window is borderless and can not be
resized as not to mess up the resolution.